jobs-for-reprocessing-tables/s3-to-s3/requirements.txt
boto3==1.19.10
botocore==1.22.10
Jinja2==3.0.2
jinjasql==0.1.8
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
